Create puzzle click handler and wire it up.
At this point the player can click puzzle pieces to move them around. However, they can click any puzzle piece to swap it with the open tile. A player should only be able to swap a puzzle piece if it is directly above or below, or directly to the left or right, of the open tile. We'll work on implementing the correct behavior next.