Add logic to only allow valid moves.
Now we only allow a player to move tiles directly above, below, to the left, or to the right of the open tile. This is the correct behavior.

Now the tiles start in the correct order, which is no challange. Next up, we need to implement code which will shuffle the tiles into a random order so there is actually a puzzle to be solved.