Add function to initialize the puzzle.
First, we do a little cleanup refactoring by creating a central function that will perform any steps necessary for making the puzzle ready for a player to solve it.

We'll add to this function in a bit, for now it should only contain the call to wire up the puzzle click handler.