Implement tile shuffling.
Now the puzzle tiles start in a random order. We have one problem however, not all random sequences of tiles are solvable in this type of puzzle. If we do not account for this it would be frustrating for any player who tries to solve an unsolvable tile sequence. Luckily, there is a formula we can use to determine if there is a solution.

See: https://www.cs.bham.ac.uk/~mdr/teaching/modules04/java2/TilesSolvability.html