Implement check for puzzle solvability.
Now the puzzle will continue to be shuffled until the tiles are in a configuration that is known to be solvable.

At this point the puzzle is playable, but once a player solves the puzzle there is no win condition notification.

The win condtion is when all tiles are in sequence, based upon the tile numbers.

Next we will implement the check for the win condtion and use a simple alert to notify the player that they have won.